+++ konsave/consts.py
@@ -1,9 +1,9 @@
 """
 This module contains all the variables for konsave
 """
+
 import os
 from konsave import __version__
-
 
 HOME = os.path.expandvars("$HOME")
 CONFIG_DIR = os.path.join(HOME, ".config")